requirements.txt
opencv-contrib-python
numpy
torch
torchvision
Pillow
